refactor(wallet): extract tRPC helpers to shared utils module
Move buildTrpcUrl, fetchTrpcData, and extractTrpcErrorMessage from
token-helpers.ts into apps/wallet/src/utils/trpc.ts so they can be
imported by any module without pulling in token-specific logic.
